Add protection for long min / long max

+++ inc/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: zaiicko <meskrabe@student.s19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 19:22:22 by zaiicko           #+#    #+#             */
-/*   Updated: 2024/05/09 03:20:09 by zaiicko          ###   ########.fr       */
+/*   Updated: 2024/05/09 03:37:33 by zaiicko          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_putchar_f(char c);
 int		ft_putstr_f(char *str);
 size_t	ft_checking(va_list *arg, char c);
 int		ft_putun_f(size_t nb, char *str);
-int		ft_putnbr_int(int nb);
+ssize_t		ft_putnbr_int(ssize_t nb);
 int		ft_putpointer(void *ptr, char *str);
 void	ft_putnbase(size_t nb, char *str);
